Update Emulator HTML
Updated HTML for QAOP so it's a bit more closed so the user can't switch games and such. Credit is still given to Jan, and the info page will stay and the HREF's will be updated to redirect to his website. Also updated to include RF1M name, and saving and loading is easier to understand for people who don't use emulators.

Removed old index.html which is no longer compatible.